test: Update tests - Include dot notation examples - Rename Tests folder to mocks
