+++ editor.planx.uk/src/@planx/components/List/schemas/mocks/Zoo.ts
@@ -25,7 +25,7 @@ export const Zoo: Schema = {
       type: "text",
       data: {
         title: "What's their email address?",
-        fn: "email",
+        fn: "email.address",
         type: TextInputType.Email,
       },
     },
@@ -57,7 +57,7 @@ export const Zoo: Schema = {
       type: "question",
       data: {
         title: "How cute are they?",
-        fn: "cuteness",
+        fn: "cuteness.amount",
         options: [
           { id: "very", data: { text: "Very" } },
           { id: "super", data: { text: "Super" } },
@@ -82,27 +82,27 @@ export const mockZooPayload = {
     mockFn: [
       {
         age: 10,
-        cuteness: "Very",
-        email: "richard.parker@pi.com",
+        "cuteness.amount": "Very",
+        "email.address": "richard.parker@pi.com",
         name: "Richard Parker",
         size: "Medium",
       },
       {
         age: 10,
-        cuteness: "Very",
-        email: "richard.parker@pi.com",
+        "cuteness.amount": "Very",
+        "email.address": "richard.parker@pi.com",
         name: "Richard Parker",
         size: "Medium",
       },
     ],
     "mockFn.one.age": 10,
-    "mockFn.one.cuteness": "Very",
-    "mockFn.one.email": "richard.parker@pi.com",
+    "mockFn.one.cuteness.amount": "Very",
+    "mockFn.one.email.address": "richard.parker@pi.com",
     "mockFn.one.name": "Richard Parker",
     "mockFn.one.size": "Medium",
     "mockFn.two.age": 10,
-    "mockFn.two.cuteness": "Very",
-    "mockFn.two.email": "richard.parker@pi.com",
+    "mockFn.two.cuteness.amount": "Very",
+    "mockFn.two.email.address": "richard.parker@pi.com",
     "mockFn.two.name": "Richard Parker",
     "mockFn.two.size": "Medium",
     "mockFn.total.listItems": 2,
